# book-network/src/main/java/com/huynh/book_network/feedbak/FeedbackResponse.java
package com.huynh.book_network.feedbak;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class FeedbackResponse {

    private Double note;
    private String comment;
    private boolean ownFeedback;
}